modularize readfile into separate segments; added more provenance
